+++ sparv/modules/korp/relations.py
@@ -18,14 +18,14 @@ MAX_POS_LENGTH = 5
 
 
 @installer("Install Korp's Word Picture SQL on remote host", language=["swe"])
-def install_relations(sqlfile: ExportInput = ExportInput("korp_wordpicture/relations.sql"),
+def install_relations(sqlfile: ExportInput = ExportInput("korp.wordpicture/relations.sql"),
                       out: OutputCommonData = OutputCommonData("korp.install_relations_marker"),
                       db_name: str = Config("korp.mysql_dbname"),
                       host: str = Config("korp.remote_host")):
     """Install Korp's Word Picture SQL on remote host.
 
     Args:
-        sqlfile (str, optional): SQL file to be installed. Defaults to ExportInput("korp_wordpicture/relations.sql").
+        sqlfile (str, optional): SQL file to be installed. Defaults to ExportInput("korp.wordpicture/relations.sql").
         out (str, optional): Marker file to be written.
         db_name (str, optional): Name of the data base. Defaults to Config("korp.mysql_dbname").
         host (str, optional): Remote host to install to. Defaults to Config("korp.remote_host").
@@ -264,7 +264,7 @@ def mi_lex(rel, x_rel_y, x_rel, rel_y):
 
 @exporter("Word Picture SQL for use in Korp", language=["swe"])
 def relations_sql(corpus: Corpus = Corpus(),
-                  out: Export = Export("korp_wordpicture/relations.sql"),
+                  out: Export = Export("korp.wordpicture/relations.sql"),
                   relations: AnnotationDataAllSourceFiles = AnnotationDataAllSourceFiles("korp.relations"),
                   source_files: Optional[AllSourceFilenames] = AllSourceFilenames(),
                   source_files_list: str = "",