+++ sparv/modules/stats_export/sbx_stats.py
@@ -71,7 +71,7 @@ def sbx_freq_list(
     lemgram: AnnotationAllSourceFiles = AnnotationAllSourceFiles("<token>:stats_export.lemgram_first"),
     complemgram: AnnotationAllSourceFiles = AnnotationAllSourceFiles(
                                             "<token>:stats_export.complemgram_best_cond"),
-    out: Export = Export("frequency_list_sbx/stats_[metadata.id].csv"),
+    out: Export = Export("stats_export.frequency_list_sbx/stats_[metadata.id].csv"),
     delimiter: str = Config("stats_export.delimiter"),
     cutoff: int = Config("stats_export.cutoff")):
     """Create a word frequency list for the entire corpus.
@@ -88,7 +88,8 @@ def sbx_freq_list(
             Defaults to AnnotationAllSourceFiles("<token>:saldo.lemgram").
         complemgram (str, optional): Conditional best compound lemgram annotations.
             Defaults to AnnotationAllSourceFiles("<token>:saldo.complemgram").
-        out (str, optional): The output word frequency file. Defaults to Export("frequency_list_sbx/[metadata.id].csv").
+        out (str, optional): The output word frequency file.
+            Defaults to Export("stats_export.frequency_list_sbx/[metadata.id].csv").
         delimiter (str, optional): Column delimiter to use in the csv. Defaults to Config("stats_export.delimiter").
         cutoff (int, optional): The minimum frequency a word must have in order to be included in the result.
             Defaults to Config("stats_export.cutoff").
@@ -107,7 +108,7 @@ def sbx_freq_list_simple_swe(
     word: AnnotationAllSourceFiles = AnnotationAllSourceFiles("<token:word>"),
     pos: AnnotationAllSourceFiles = AnnotationAllSourceFiles("<token:pos>"),
     baseform: AnnotationAllSourceFiles = AnnotationAllSourceFiles("<token>:stats_export.baseform_first"),
-    out: Export = Export("frequency_list_sbx/stats_[metadata.id].csv"),
+    out: Export = Export("stats_export.frequency_list_sbx/stats_[metadata.id].csv"),
     delimiter: str = Config("stats_export.delimiter"),
     cutoff: int = Config("stats_export.cutoff")):
     """Create a word frequency list for a corpus without sense, lemgram and complemgram annotations."""
@@ -124,7 +125,7 @@ def sbx_freq_list_simple(
     word: AnnotationAllSourceFiles = AnnotationAllSourceFiles("<token:word>"),
     pos: AnnotationAllSourceFiles = AnnotationAllSourceFiles("<token:pos>"),
     baseform: AnnotationAllSourceFiles = AnnotationAllSourceFiles("<token:baseform>"),
-    out: Export = Export("frequency_list_sbx/stats_[metadata.id].csv"),
+    out: Export = Export("stats_export.frequency_list_sbx/stats_[metadata.id].csv"),
     delimiter: str = Config("stats_export.delimiter"),
     cutoff: int = Config("stats_export.cutoff")):
     """Create a word frequency list for a corpus without sense, lemgram and complemgram annotations."""
@@ -141,7 +142,7 @@ def sbx_freq_list_fsv(
     word: AnnotationAllSourceFiles = AnnotationAllSourceFiles("<token:word>"),
     baseform: AnnotationAllSourceFiles = AnnotationAllSourceFiles("<token:baseform>"),
     lemgram: AnnotationAllSourceFiles = AnnotationAllSourceFiles("<token:lemgram>"),
-    out: Export = Export("frequency_list_sbx/stats_[metadata.id].csv"),
+    out: Export = Export("stats_export.frequency_list_sbx/stats_[metadata.id].csv"),
     delimiter: str = Config("stats_export.delimiter"),
     cutoff: int = Config("stats_export.cutoff")):
     """Create a word frequency list for a corpus without sense, lemgram and complemgram annotations."""
@@ -153,7 +154,7 @@ def sbx_freq_list_fsv(
 
 @installer("Install SBX word frequency list on remote host")
 def install_sbx_freq_list(
-    freq_list: ExportInput = ExportInput("frequency_list_sbx/stats_[metadata.id].csv"),
+    freq_list: ExportInput = ExportInput("stats_export.frequency_list_sbx/stats_[metadata.id].csv"),
     out: OutputCommonData = OutputCommonData("stats_export.install_sbx_freq_list_marker"),
     host: str = Config("stats_export.remote_host"),
     target_dir: str = Config("stats_export.remote_dir")):